pyproject.toml
[tool.poetry]
name = "happy-hub"
version = "0.1.0"
description = ""
authors = ["thanglq150188 <thanglq150188@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.42.0"
sentence-transformers = "^3.0.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
jsonschema = "^4.23.0"
docstring-parser = "^0.16"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
